233 bug 修复，修复溢出问题

+++ src/com/douma/_5_day/_233/_233_number_of_digit_one.java
@@ -11,7 +11,8 @@ public class _233_number_of_digit_one {
     // O(logn)
     public int countDigitOne(int n) {
         int count = 0;
-        for (int i = 1; i <= n; i *= 10) {
+        // bug 修复：i 应该是长整型，要不然会溢出
+        for (long i = 1; i <= n; i *= 10) {
             // (n/10)*1 + min(max((n%10 - 1 + 1), 0), 1)
             // (n/100)*10 + min(max((n%100 - 10 + 1), 0), 10)
             // (n/1000)*100 + min(max((n%1000 - 100 + 1), 0), 100)
